pyproject.toml
[project]
name = "circuit-playground-pro"
version = "3.1.0"
description = "Professional-grade visual logic circuit simulator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.10.1",
    "PyYAML>=6.0.2",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.17.0",
    "nuitka>=2.8.9",
    "zstandard>=0.25.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-qt>=4.4.0",
]